backend/image-service/requirements.txt
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework-simplejwt==5.4.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.0
Pillow==10.4.0
pyvips==2.2.3
python-dotenv==1.0.1
python-magic==0.4.27
drf-yasg==1.21.7
